This might be an issue, casting a float for now

+++ dev/PID.c
@@ -47,5 +47,5 @@ float PID_computeOffset(PID *pid, float sensorValue) {
 }
 
 void PID_resetPIDerror(PID* pid, float4 error){
-    pid->totalError = error;
+    pid->totalError = (float)(error);
 }
